+++ src/lib/AsyncTypeaheadInput.tsx
@@ -195,7 +195,9 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadInputPr
           const values = convertAutoCompleteOptionsToStringArray(optionsArray);
           const finalValue = multiple ? values : values[0];
           if (onChange) {
-            onChange(finalValue);
+            void (async () => {
+              await onChange(finalValue);
+            })();
           }
           field.onChange(finalValue);
         }}