+++ src/lib/types/Typeahead.ts
@@ -29,7 +29,7 @@ interface CommonTypeaheadProps<T extends FieldValues> extends Omit<
   innerRef?: RefObject<HTMLInputElement | null>;
   fitMenuContent?: boolean;
   getOptionDisabled?: (option: TypeaheadOption) => boolean;
-  onChange?: (selected: string | string[]) => void;
+  onChange?: (selected: string | string[]) => void | Promise<void>;
   onInputChange?: (text: string, reason: AutocompleteInputChangeReason) => void;
   onClose?: (event: SyntheticEvent, reason: AutocompleteCloseReason) => void;
   onOpen?: (event: SyntheticEvent) => void;